requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
google-auth==2.6.6
google-auth-oauthlib==0.5.1
gspread==5.3.2
idna==3.3
numpy==1.21.6
oauthlib==3.2.0
pandas==1.3.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
